Migrate most-liked sync to a generic integrations sync pipeline
