+++ orchestration/sync_pipeline.py
@@ -3,8 +3,6 @@
 import os
 
 from prefect import task, flow
-from prefect.deployments import Deployment
-from prefect.server.schemas.schedules import CronSchedule
 
 from orchestration.helper import pipelines_directory, run_slurm_job
 
@@ -36,32 +34,11 @@ def write_firehose_data():
     return result
 
 
-@task(log_prints=True)
-def sync_most_liked():
-    """Syncs the most liked posts."""
-    bash_script_path = os.path.join(
-        pipelines_directory, "sync_post_records", "most_liked", "submit_job.sh"
-    )
-    result = run_slurm_job(bash_script_path, timeout_ok=False)
-    if result is None:
-        raise Exception("SLURM job failed in sync_most_liked task")
-    return result
-
-
 @flow(name="Sync pipeline", log_prints=True)
-def sync_data_pipeline(run_most_liked=True):
+def sync_data_pipeline():
     """Syncs the data pipeline."""
     sync_firehose.submit(wait_for=False)
     write_firehose_data.submit(wait_for=False)
-
-    if run_most_liked:
-        sync_most_liked.submit(wait_for=False)
-
-
-@flow(name="Most Liked Sync", log_prints=True)
-def sync_most_liked_pipeline():
-    """Syncs only the most liked posts."""
-    sync_most_liked.submit()
 
 
 if __name__ == "__main__":
@@ -75,17 +52,6 @@ if __name__ == "__main__":
     #     tags=["slurm", "prod", "firehose"],
     #     parameters={"run_most_liked": False},
     # )
-
-    most_liked_deployment = Deployment.build_from_flow(
-        flow=sync_most_liked_pipeline,
-        name="Most Liked Sync Pipeline",
-        schedule=CronSchedule(cron="0 */4 * * *"),
-        tags=["slurm", "prod", "most_liked"],
-    )
-
-    # Apply deployments
-    # daily_sync_deployment.apply()
-    most_liked_deployment.apply()
 
     # Run the initial sync_data_pipeline
     sync_data_pipeline(run_most_liked=True)
